requirements.txt
PyTelegramBotAPI
piexif
pillow